+++ package.json
@@ -7,8 +7,9 @@
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
     "@mui/icons-material": "^5.6.2",
-    "@mui/material": "^5.6.3",
+    "@mui/material": "^5.6.4",
     "@mui/styled-engine-sc": "^5.6.1",
+    "@mui/styles": "^5.6.2",
     "@reduxjs/toolkit": "^1.8.1",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.1.1",
