+++ src/components/Header/Header.tsx
@@ -1,17 +1,16 @@
 /* eslint-disable */
-import React from "react";
+import { AppBar } from "@mui/material";
+import { Theme } from "@mui/material/styles";
+import { makeStyles } from "@mui/styles";
 import clsx from "clsx";
-import { makeStyles } from "@material-ui/core/styles";
-
-import AppBar from "@material-ui/core/AppBar";
-
-import Toggle from "../Svg/Toggle";
-import "./Header.scss";
+import React from "react";
 import FeedBack from "../Svg/FeedBack";
 import Language from "../Svg/Language";
 import ThemeMode from "../Svg/ThemeMode";
+import Toggle from "../Svg/Toggle";
+import "./Header.scss";
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles((theme: Theme) => ({
   appBar: {
     zIndex: theme.zIndex.drawer + 1,
     transition: theme.transitions.create(["margin", "width"], {
@@ -47,7 +46,7 @@ const Header = ({ open, setOpen }: Props) => {
 
   return (
     <AppBar
-    elevation={0}
+      elevation={0}
       position="fixed"
       className={clsx(classes.appBar, {
         [classes.appBarShift]: open,
